# Make Allow middleware thread-safe without duping