--- a/lib/strong_routes.rb
+++ b/lib/strong_routes.rb
@@ -1,4 +1,4 @@
-require "rack/request"
+require "rack/response"
 
 require "strong_routes/allow"
 require "strong_routes/config"
@@ -6,8 +6,16 @@ require "strong_routes/route_matcher"
 require "strong_routes/version"
 
 module StrongRoutes
+  def self.allowed?(route)
+    config.route_matchers.any? { |route_matcher| route_matcher =~ route }
+  end
+
   def self.config
     @config ||= Config.new
+  end
+
+  def self.enabled?
+    config.enabled?
   end
 
   # Initialize the config