The initial `Allow` middleware implementation [was not thread-safe](https://github.com/liveh2o/strong_routes/commit/61a3b507e83f6a582d1cee9f36b70306b0fc0eae). A "temporary" (eight years ago!) hack was put in place that simply duped the middleware stack to ensure threadsafety. Not only is this a hack, but it also adds unnecessary overhead to each request since each dup generated more objects that would need to be GC'd.

This change makes the `Allow` middleware thread-safe without duping, and yields a whopping 85x (!) increase in performance (via bin/benchmark):

Before:

    ruby 2.7.6p219 (2022-04-12 revision c9c2245c0a) [arm64-darwin22]
    Warming up --------------------------------------
        Hit    2.318k i/100ms
        Miss   2.317k i/100ms
    Calculating -------------------------------------
        Hit   22.267k (± 2.8%) i/s (44.91 μs/i) - 111.264k in 5.000717s
        Miss  22.266k (± 1.2%) i/s (44.91 μs/i) - 113.533k in 5.099622s

After:

    ruby 2.7.6p219 (2022-04-12 revision c9c2245c0a) [arm64-darwin22]
    Warming up --------------------------------------
        Hit  188.921k i/100ms
        Miss 186.268k i/100ms
    Calculating -------------------------------------
        Hit    1.859M (± 0.5%) i/s (537.92 ns/i) - 9.446M in 5.081332s
        Miss   1.860M (± 1.1%) i/s (537.70 ns/i) - 9.313M in 5.008448s

This implementation uses a simplified Config class that is no longer backed by a hash. This required dropping support for passing config options to the middleware, but that likely was mostly used in tests.